models/TST: Docstring fixup on MHA layers
Fix documentation on default value of d_k and d_v. If they are not provided, and
d_model and n_heads are kept at default values, they will be set to 128//16
which is 8.

Also, specify the usual values range to 8-64 which should more or less correspond
to a d_model/n_heads range.